# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "0.0.1",
    "title": "Winding tree API",
    "description": "Node js API for Winding Tree",
    "contact": {
      "name": "Winding tree",
      "url": "https://windingtree.com/"
    },
    "license": {
      "name": "Apache-2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0"
    }
  },
...
  "schemes": [
    "http"
  ],
  "host": "localhost:3000",
  "basePath": "/",
...
  "paths": {
    "/balance": {
      "get": {
        "tags": [ "User" ],
        "summary": "Get balance.",
        "description": "Returns true if the account have more Líf tokens",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Body.",
            "required": true,
            "schema": {
              "required": [
                "account",
                "cost"
              ],
              "properties": {
                "account": {
                  "description": "Account to verify balance.",
                  "type": "string"
                },
                "cost": {
                  "description": "mount of Líf to verify.",
                  "type": "number"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ok.",
            "schema": {
              "balance":{
                "type": "boolean"
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
...
    "/hotels/{hotelAddress}/units/{unitAddress}/defaultPrice" : {
      "put": {
        "tags": [ "Units" ],
        "summary": "Set default price.",
        "description": "Set default price for a unit of hotel.",
        "parameters": [
          {
            "in": "path",
            "name": "address",
            "type": "string",
            "required": true,
            "description": "Hotel address."
          },
          {
            "in": "path",
            "name": "unit",
            "type": "string",
            "required": true,
            "description": "unit address."
          },
          {
            "in": "body",
            "name": "params",
            "required": true,
            "description": "parameters.",
            "schema": {
              "required": [
                "password",
                "price"
              ],
              "properties": {
                "password": {
                  "description": "Current password.",
                  "type": "string"
                },
                "price": {
                  "description": "Default price.",
                  "type": "number"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sucefully updated.",
            "schema": {
              "$ref": "#/definitions/txHash"
            }
          },
          "400": {
            "description": "Invalid input",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/hotels/{hotelAddress}/units/{unitAddress}/defaultLifPrice" : {
      "post": {
        "tags": [ "Units" ],
        "summary": "Set default Lif price.",
        "description": "Set default Lif price for a unit of hotel.",
        "parameters": [
          {
            "in": "path",
            "name": "address",
            "type": "string",
            "required": true,
            "description": "Hotel address."
          },
          {
            "in": "path",
            "name": "unit",
            "type": "string",
            "required": true,
            "description": "unit address."
          },
          {
            "in": "body",
            "name": "params",
            "required": true,
            "description": "parameters.",
            "schema": {
              "required": [
                "password",
                "price"
              ],
              "properties": {
                "password": {
                  "description": "Current password.",
                  "type": "string"
                },
                "price": {
                  "description": "Default price.",
                  "type": "number"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sucefully updated.",
            "schema": {
              "$ref": "#/definitions/txHash"
            }
          },
          "400": {
            "description": "Invalid input",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/hotels/{hotelAddress}/units/{unitAddress}/specialLifPrice" : {
      "post": {
        "tags": [ "Units" ],
        "summary": "Set special Lif price.",
        "description": "Set special Lif price for a unit of hotel.",
        "parameters": [
          {
            "in": "path",
            "name": "address",
            "type": "string",
            "required": true,
            "description": "Hotel address."
          },
          {
            "in": "path",
            "name": "unit",
            "type": "string",
            "required": true,
            "description": "unit address."
          },
          {
            "in": "body",
            "name": "params",
            "required": true,
            "description": "parameters.",
            "schema": {
              "required": [
                "password",
                "price",
                "from",
                "days"
              ],
              "properties": {
                "password": {
                  "description": "Current password.",
                  "type": "string"
                },
                "price": {
                  "description": "Special price.",
                  "type": "number"
                },
                "from": {
                  "description": "Initla date.",
                  "type": "Date"
                },
                "dats": {
                  "description": "Amount of days.",
                  "type": "number"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sucefully updated.",
            "schema": {
              "$ref": "#/definitions/txHash"
            }
          },
          "400": {
            "description": "Invalid input",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/hotels/{hotelAddress}/units/{unitAddress}/specialPrice" : {
      "post": {
        "tags": [ "Units" ],
        "summary": "Set special price.",
        "description": "Set special price for a unit of hotel.",
        "parameters": [
          {
            "in": "path",
            "name": "address",
            "type": "string",
            "required": true,
            "description": "Hotel address."
          },
          {
            "in": "path",
            "name": "unit",
            "type": "string",
            "required": true,
            "description": "unit address."
          },
          {
            "in": "body",
            "name": "params",
            "required": true,
            "description": "parameters.",
            "schema": {
              "required": [
                "password",
                "price",
                "from",
                "days"
              ],
              "properties": {
                "password": {
                  "description": "Current password.",
                  "type": "string"
                },
                "price": {
                  "description": "Special price.",
                  "type": "number"
                },
                "from": {
                  "description": "Initla date.",
                  "type": "Date"
                },
                "dats": {
                  "description": "Amount of days.",
                  "type": "number"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sucefully updated.",
            "schema": {
              "$ref": "#/definitions/txHash"
            }
          },
          "400": {
            "description": "Invalid input",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/hotels/{hotelAddress}/units/{unitAddress}/currencyCode" : {
       "post": {
         "tags": [ "Units" ],
         "summary": "Set currency Code.",
         "description": "Set currency code for a unit of hotel.",
         "parameters": [
           {
             "in": "path",
             "name": "address",
             "type": "string",
             "required": true,
             "description": "Hotel address."
           },
           {
             "in": "path",
             "name": "unit",
             "type": "string",
             "required": true,
             "description": "unit address."
           },
           {
             "in": "body",
             "name": "params",
             "required": true,
             "description": "parameters.",
             "schema": {
               "required": [
                 "password",
                 "code"
               ],
               "properties": {
                 "password": {
                   "description": "Current password.",
                   "type": "string"
                 },
                 "code": {
                   "description": "Currency code.",
                   "type": "number"
                 }
               }
             }
           }
         ],
         "responses": {
           "200": {
             "description": "Sucefully updated.",
             "schema": {
               "$ref": "#/definitions/txHash"
             }
           },
           "400": {
             "description": "Invalid input",
             "schema": {
               "$ref": "#/definitions/Error"
             }
           }
         }
       }
     },
...
    "/units/{unitAddress}/cost":{
        "get":{
          "tags": [ "Units" ],
          "summary": "Get cost",
          "description": "Get cost for an unit.",
          "parameters": [
            {
              "in": "path",
              "name": "unit",
              "type": "string",
              "required": true,
              "description": "Unit address."
            },
            {
              "name": "body",
              "in": "body",
              "schema": {
                "required": [
                  "from",
                  "days"
                ],
                "properties": {
                  "from": {
                    "description": "Start date. UTC day since 1970",
                    "type": "number"
                  },
                  "days": {
                    "description": "Amount of days of the reservation.",
                    "type": "numer"
                  }
                }
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Ok.",
              "schema": {
                "$ref": "#/definitions/Reservation"
              }
            },
            "400": {
              "description": "Can not get.",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        }
      },
      "/units/{unitAddress}/lifCost":{
          "get":{
            "tags": [ "Units" ],
            "summary": "Get Lif costs.",
            "description": "Get Lif costs for an unit.",
            "parameters": [
              {
                "in": "path",
                "name": "unit",
                "type": "string",
                "required": true,
                "description": "Unit address."
              },
              {
                "name": "body",
                "in": "body",
                "schema": {
                  "required": [
                    "from",
                    "days"
                  ],
                  "properties": {
                    "from": {
                      "description": "Start date. UTC day since 1970",
                      "type": "number"
                    },
                    "days": {
                      "description": "Amount of days of the period.",
                      "type": "numer"
                    }
                  }
                }
              }
            ],
            "responses": {
              "200": {
                "description": "Ok.",
                "schema": {
                  "$ref": "#/definitions/Reservation"
                }
              },
              "400": {
                "description": "Can not get.",
                "schema": {
                  "$ref": "#/definitions/Error"
                }
              }
            }
          }
      },
...
  },
  "definitions": {
    "txHash": {
      "title": "Transaction hash",
      "description": "transaction hash for the operation.",
      "properties": {
        "txHash": {
          "type": "string",
          "description": "Transaction hash."
        }
      }
    },
    "Error": {
      "title": "Error",
      "description": "Default schema for errors returned by API.",
      "properties": {
        "code": {
          "type": "string",
          "description": "Computer-readable error. Use this for comparison."
        },
        "short": {
          "type": "string",
          "description": "Human-readable error with a short description of the error."
        },
        "long": {
          "type": "string",
          "description": "Verbose explanation of what happen."
        }
      }
    },
...
    "Reservation": {
      "title": "Reservation",
      "description": "Reservation schema .",
      "properties": {
        "specialPrice": {
          "type": "string",
          "description": "Specil price for reservation."
        },
        "specialLifPrice": {
          "type": "string",
          "description": "Specil Lif price for reservation."
        },
        "BookedBy": {
          "type": "string",
          "description": "Addres that booked the unit."
        }
      }
    },
...
  }
}